Pass double from GUI as stream rate instead of int. Handle rate==0 by sending stopStream() instead of setting rate to 0. Tweaked gl blending settings to no real effect.

+++ src/clients/legacyWatcher/manetglview.h
@@ -50,7 +50,7 @@ public slots:
         void forwardPlayback();
         void rewindToStartOfPlayback();
         void forwardToEndOfPlayback();
-        void playbackSetSpeed(int speed);
+        void playbackSetSpeed(double speed);
 
         void toggleNodeSelectedForGraph(unsigned int nodeId);
         void showNodeSelectedForGraph(unsigned int nodeId, bool);
